--- src/components/cardList/Loading/Loading.stories.tsx
@@ -1,15 +1,0 @@
-import type { Meta, StoryObj } from '@storybook/react-vite';
-import { Loading } from './Loading';
-
-const meta: Meta<typeof Loading> = {
-  title: 'Components/CardList/Loading',
-  component: Loading,
-};
-
-export default meta;
-
-type Story = StoryObj<typeof Loading>;
-
-export const Default: Story = {
-  render: () => <Loading />,
-};
